refactor: re-structuring files
