+++ practice-two/src/components/Loader/index.tsx
