+++ practice-two/src/pages/StudentPage/hooks/useStudentRemoving.tsx
@@ -1,10 +1,11 @@
 import { useMutation, useQueryClient } from 'react-query';
-import api from '@services/apiRequest';
+import api from '@services/api-request';
 import { STUDENTS_URL } from '@constants/services';
 import { TStudent } from 'src/types';
 
 const useStudentRemoving = () => {
   const queryClient = useQueryClient();
+
   const { mutateAsync, mutate } = useMutation({
     mutationFn: async (id: string | number) =>
       (await api.remove(STUDENTS_URL + '/' + id)) as TStudent,
